fix errors with yaml file location and parse errors

+++ ros2_ws/src/sensors/include/sensors/depth_calibration.hpp
@@ -1,5 +1,5 @@
 namespace sensors
 {
 void load_depth_configuration();
-float get_calibrated_depth(float base_depth);
+double get_calibrated_depth(double base_depth);
 }
